refactoring and testing: added tests for debug with selector, refactored DebugCommand and execute-client-function
