+++ src/client-functions/client-function-builder.js
@@ -2,7 +2,7 @@ import { isNil as isNullOrUndefined, assign } from 'lodash';
 import testRunTracker from '../api/test-run-tracker';
 import functionBuilderSymbol from './builder-symbol';
 import { createReplicator, FunctionTransform } from './replicator';
-import { ExecuteClientFunctionCommand } from '../test-run/commands/observation';
+import { ExecuteClientFunctionCommand } from '../test-run/commands/execute-client-function';
 import compileClientFunction from '../compiler/compile-client-function';
 import { APIError, ClientFunctionAPIError } from '../errors/runtime';
 import { assertType, is } from '../errors/runtime/type-assertions';
