+++ src/test-run/commands/execute-client-function.d.ts
@@ -1,0 +1,24 @@
+import { CommandBase } from './base';
+import TestRun from '../index';
+
+declare class ExecuteClientFunctionCommandBase extends CommandBase {
+    public constructor(obj: object, testRun: TestRun, type: string);
+    public instantiationCallsiteName: string;
+    public fnCode: string;
+    public args: string[];
+    public dependencies: string[];
+}
+
+export class ExecuteClientFunctionCommand extends ExecuteClientFunctionCommandBase {
+    public constructor(obj: object, testRun: TestRun);
+}
+
+export class ExecuteSelectorCommand extends ExecuteClientFunctionCommandBase {
+    public constructor(obj: object, testRun: TestRun);
+    public visibilityCheck: boolean;
+    public timeout?: number;
+    public apiFnChain: string[];
+    public needError: boolean;
+    public index: number;
+    public strictError: boolean;
+}
