+++ src/test-run/commands/observation.js
@@ -2,6 +2,9 @@ import TYPE from './type';
 import { ActionCommandBase } from './base';
 import { positiveIntegerArgument } from './validations/argument';
 import { camelCase } from 'lodash';
+import {
+    initSelector,
+} from './validations/initializers';
 
 // Commands
 export class WaitCommand extends ActionCommandBase {
@@ -18,52 +21,20 @@ export class WaitCommand extends ActionCommandBase {
     }
 }
 
-export class ExecuteClientFunctionCommandBase extends ActionCommandBase {
-    constructor (obj, testRun, type) {
-        super(obj, testRun, type, false);
-    }
-
-    getAssignableProperties () {
-        return [
-            { name: 'instantiationCallsiteName', defaultValue: '' },
-            { name: 'fnCode', defaultValue: '' },
-            { name: 'args', defaultValue: [] },
-            { name: 'dependencies', defaultValue: [] },
-        ];
-    }
-}
-
-export class ExecuteClientFunctionCommand extends ExecuteClientFunctionCommandBase {
-    static methodName = TYPE.executeClientFunction;
-
-    constructor (obj, testRun) {
-        super(obj, testRun, TYPE.executeClientFunction);
-    }
-}
-
-export class ExecuteSelectorCommand extends ExecuteClientFunctionCommandBase {
-    static methodName = TYPE.executeSelector;
-
-    constructor (obj, testRun) {
-        super(obj, testRun, TYPE.executeSelector);
-    }
-
-    getAssignableProperties () {
-        return [
-            { name: 'visibilityCheck', defaultValue: false },
-            { name: 'timeout', defaultValue: null },
-            { name: 'apiFnChain' },
-            { name: 'needError' },
-            { name: 'index', defaultValue: 0 },
-            { name: 'strictError' },
-        ];
-    }
-}
-
 export class DebugCommand extends ActionCommandBase {
     static methodName = camelCase(TYPE.debug);
 
-    constructor () {
-        super(null, null, TYPE.debug);
+    constructor (obj, testRun,) {
+        super(obj, testRun, TYPE.debug);
+    }
+
+    getAssignableProperties () {
+        return [
+            { name: 'selector', init: (name, val, options) => {
+                return initSelector(name, val, Object.assign({}, options,
+                    { skipVisibilityCheck: true, collectionMode: true }
+                ));
+            }, required: false },
+        ];
     }
 }
